requirements.txt
fastapi
uvicorn  
tensorflow
numpy
Pillow 
python-multipart
opencv-python
scikit-learn
matplotlib
